pyproject.toml
[project]
name = "canvas-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "canvasapi>=3.5.0",
    "python-dotenv>=1.2.2",
]
